refactor into single global state
