+++ frontend/package.json.md5
@@ -1,1 +1,1 @@
-3ad5118eed28c2edaeb0758dd2ea3532
+d2356f779e43b19c48c4bafee6f9b00d
