+++ frontend/src/contexts/global/global-actions.ts
@@ -1,30 +1,48 @@
 import { requestPassphrase } from '../../components/passphrase-modal'
 import { AppToaster } from '../../components/toaster'
+import { DataSources } from '../../config/data-sources'
 import { DEFAULT_VEGA_HOME, IS_TEST_MODE } from '../../config/environment'
 import { Intent } from '../../config/intent'
 import { createLogger } from '../../lib/logging'
 import { Service } from '../../service'
-import type { FirstPublicKey } from '../../wailsjs/go/models'
+import type {
+  Config,
+  FirstPublicKey,
+  GetServiceStateResponse,
+  Network,
+  StartServiceRequest
+} from '../../wailsjs/go/models'
 import { GenerateKeyRequest } from '../../wailsjs/go/models'
 import type { GlobalDispatch, GlobalState } from './global-context'
+import { ProxyName } from './global-context'
 import type { GlobalAction } from './global-reducer'
 
 const logger = createLogger('GlobalActions')
 
 export function initAppAction() {
   return async (dispatch: GlobalDispatch) => {
+    let isInit = false
+    let appConfig: Config | null = null
+    let defaultNetwork = null
+    let defaultNetworkConfig = null
+
     try {
-      const config = await Service.GetAppConfig()
-      dispatch({ type: 'SET_CONFIG', config: config })
+      isInit = await Service.IsAppInitialised()
     } catch (err) {
       logger.error(err)
     }
 
+    if (isInit) {
+      try {
+        appConfig = await Service.GetAppConfig()
+        dispatch({ type: 'SET_CONFIG', config: appConfig })
+      } catch (err) {
+        console.log('failed to get config', err)
+      }
+    }
+
     try {
-      const [isInit, version] = await Promise.all([
-        Service.IsAppInitialised(),
-        Service.GetVersion()
-      ])
+      const version = await Service.GetVersion()
 
       dispatch({ type: 'SET_VERSION', version: version.version })
 
@@ -48,6 +66,95 @@ export function initAppAction() {
       dispatch({ type: 'INIT_APP', isInit: true })
     } catch (err) {
       dispatch({ type: 'INIT_APP', isInit: false })
+      logger.error(err)
+    }
+
+    try {
+      const res = await fetch(DataSources.NETWORKS)
+      const json = await res.json()
+      dispatch({ type: 'SET_PRESETS', presets: json })
+    } catch (err) {
+      logger.error(err)
+    }
+
+    try {
+      const networks = await Service.ListNetworks()
+      if (networks instanceof Error) throw networks
+
+      if (networks.networks.length) {
+        defaultNetwork = networks.networks[0]
+
+        if (appConfig?.defaultNetwork) {
+          defaultNetwork =
+            networks.networks.find(n => n === appConfig?.defaultNetwork) ||
+            networks.networks[0]
+        }
+
+        defaultNetworkConfig = await Service.GetNetworkConfig(defaultNetwork)
+
+        dispatch({
+          type: 'SET_NETWORKS',
+          network: defaultNetwork,
+          networks: networks.networks,
+          config: defaultNetworkConfig
+        })
+      } else {
+        dispatch({
+          type: 'SET_NETWORKS',
+          network: null,
+          networks: [],
+          config: null
+        })
+      }
+    } catch (err) {
+      logger.error(err)
+    }
+
+    /**
+     * Set proxy state on startup. When running the actual app proxies will never be running at
+     * startup but if using a browser this ensures proxy state is correct if you reload the window
+     * without stopping the backend process.
+     */
+    try {
+      const [consoleState, tokenDappState] = await Promise.all([
+        Service.GetConsoleState(),
+        Service.GetTokenDAppState()
+      ])
+
+      if (consoleState.running) {
+        dispatch({
+          type: 'START_PROXY',
+          app: ProxyName.Console,
+          url: consoleState.url
+        })
+      }
+
+      if (tokenDappState.running) {
+        dispatch({
+          type: 'START_PROXY',
+          app: ProxyName.TokenDApp,
+          url: tokenDappState.url
+        })
+      }
+    } catch (err) {
+      logger.error(err)
+    }
+
+    logger.debug('StartService')
+    try {
+      const status = await Service.GetServiceState()
+
+      if (status.running) {
+        await Service.StopService()
+      }
+
+      if (defaultNetwork && defaultNetworkConfig) {
+        await Service.StartService({
+          network: defaultNetwork
+        })
+        dispatch({ type: 'START_SERVICE', port: defaultNetworkConfig.port })
+      }
+    } catch (err) {
       logger.error(err)
     }
   }
@@ -149,5 +256,205 @@ export function deactivateWalletAction(wallet: string): GlobalAction {
   return {
     type: 'DEACTIVATE_WALLET',
     wallet
+  }
+}
+
+// Network actions
+
+export function changeNetworkAction(network: string) {
+  return async (dispatch: GlobalDispatch, getState: () => GlobalState) => {
+    const state = getState()
+
+    logger.debug('ChangeNetwork')
+
+    try {
+      await stopProxies()
+      dispatch({ type: 'STOP_ALL_PROXIES' })
+
+      console.log(state.config)
+      // Save selected network to app config
+      const logLevel = state.config
+        ? state.config.logLevel.charAt(0).toUpperCase() +
+          state.config.logLevel.slice(1)
+        : 'Info'
+      await Service.UpdateAppConfig({
+        ...state.config,
+        logLevel,
+        defaultNetwork: network
+      })
+
+      const config = await Service.GetNetworkConfig(network)
+
+      dispatch({
+        type: 'CHANGE_NETWORK',
+        network,
+        config
+      })
+    } catch (err) {
+      AppToaster.show({
+        message: `${err}`,
+        intent: Intent.DANGER
+      })
+      logger.error(err)
+    }
+  }
+}
+
+export function updateNetworkConfigAction(
+  editingNetwork: string,
+  config: Network
+) {
+  return async (dispatch: GlobalDispatch, getState: () => GlobalState) => {
+    const state = getState()
+
+    logger.debug('UpdateNetworkConfig')
+
+    try {
+      // Stop main REST service if you are editing the active network config
+      if (state.network === editingNetwork) {
+        const serviceStatus = await Service.GetServiceState()
+        if (serviceStatus.running) {
+          await Service.StopService()
+        }
+      }
+
+      await stopProxies()
+      dispatch({ type: 'STOP_ALL_PROXIES' })
+
+      const isSuccessful = await Service.SaveNetworkConfig(config)
+
+      if (isSuccessful) {
+        AppToaster.show({
+          message: 'Configuration saved',
+          intent: Intent.SUCCESS
+        })
+        dispatch({ type: 'UPDATE_NETWORK_CONFIG', config })
+      } else {
+        AppToaster.show({ message: 'Error: Unknown', intent: Intent.DANGER })
+      }
+
+      if (!state.network) {
+        throw new Error('No network selected')
+      }
+
+      await Service.StartService({
+        network: state.network
+      })
+    } catch (err) {
+      AppToaster.show({ message: `${err}`, intent: Intent.DANGER })
+      logger.error(err)
+    }
+  }
+}
+
+export function addNetworkAction(
+  network: string,
+  config: Network
+): GlobalAction {
+  return {
+    type: 'ADD_NETWORK',
+    network,
+    config
+  }
+}
+
+export function stopServiceAction() {
+  return async (dispatch: GlobalDispatch) => {
+    logger.debug('StopService')
+    try {
+      const status = await Service.GetServiceState()
+      if (status.running) {
+        await Service.StopService()
+        dispatch({ type: 'STOP_SERVICE' })
+      }
+    } catch (err) {
+      logger.error(err)
+    }
+  }
+}
+
+export function startProxyAction(
+  network: string,
+  proxyAppName: ProxyName,
+  url: string
+) {
+  const proxyFns = ProxyFns[proxyAppName]
+
+  return async (dispatch: GlobalDispatch) => {
+    logger.debug('StartProxy:', proxyAppName, network)
+    try {
+      const status = await proxyFns.GetState()
+
+      if (status.running) {
+        await proxyFns.Stop()
+      }
+
+      dispatch({
+        type: 'START_PROXY',
+        app: proxyAppName,
+        url
+      })
+
+      await proxyFns.Start({
+        network
+      })
+    } catch (err) {
+      logger.error(err)
+    }
+  }
+}
+
+export function stopProxyAction(proxyAppName: ProxyName) {
+  const proxyFns = ProxyFns[proxyAppName]
+
+  return async (dispatch: GlobalDispatch) => {
+    try {
+      const status = await proxyFns.GetState()
+
+      if (status.running) {
+        await proxyFns.Stop()
+      }
+
+      dispatch({ type: 'STOP_PROXY', app: proxyAppName })
+    } catch (err) {
+      logger.error(err)
+    }
+  }
+}
+
+const ProxyFns: {
+  [A in ProxyName]: {
+    GetState: () => Promise<GetServiceStateResponse>
+    Start: (req: StartServiceRequest) => Promise<boolean | Error>
+    Stop: () => Promise<boolean | Error>
+  }
+} = {
+  [ProxyName.Console]: {
+    GetState: Service.GetConsoleState,
+    Start: Service.StartConsole,
+    Stop: Service.StopConsole
+  },
+  [ProxyName.TokenDApp]: {
+    GetState: Service.GetTokenDAppState,
+    Start: Service.StartTokenDApp,
+    Stop: Service.StopTokenDApp
+  }
+}
+
+async function stopProxies() {
+  try {
+    // Stop Console
+    const consoleStatus = await Service.GetConsoleState()
+    if (consoleStatus.running) {
+      await Service.StopConsole()
+    }
+
+    // Stop TokenDapp
+    const tokenDappStatus = await Service.GetTokenDAppState()
+    if (tokenDappStatus.running) {
+      await Service.StopTokenDApp()
+    }
+  } catch (err) {
+    logger.error(err)
   }
 }
